Changed: #99: Check to use buffer streaming or music channel depending on driver implementation. Changed: #29: Fixed some compiler warnings.
git-svn-id: http://dev.ryzom.com/svn/trunk@1091 60ea55b9-4480-4529-b8ac-ed5291525b0f

+++ nel/src/sound/driver/xaudio2/music_buffer_vorbis.cpp
@@ -178,7 +178,7 @@ uint32 CMusicBufferVorbis::getNextBytes(uint8 *buffer, uint32 minimum, uint32 ma
 uint16 CMusicBufferVorbis::getChannels()
 {
 	vorbis_info *vi = ov_info(&_OggVorbisFile, -1);
-	return vi->channels;
+	return (uint16)vi->channels;
 }
 
 uint32 CMusicBufferVorbis::getSamplesPerSec()
